Change versioning
Change the program version to 0.00.dev - this will be updated to a "real"
version when preparing a release.
Add an internal file_version (set to 3000 as a starting value). This will
be incremented whenever a change occurs that affects backwards
compatibility of the file structure.

Added a GetVersionString() function so the logic for generating
a version string is isolated to one place.