ref:template:#83 Change keeps_list
Move the Ttemplate class into a unit of it's own, along with all the
records it depends on.

Pull the Tmark record out into a unit of it's own.

Change the keeps_list from a TStringList to a TTemplateList, and remove
all the typecasts.

Fix up the sorting code in keep_select.