feat:t2box: Add initial (inactive) menu entry
Apart from changes too hook this facility into the existing menus,
this commit contains a dummy skeleton for t2unit which is based on a
stripped down version of mecbox_unit.pas.

The next version will replaced this with a copy of the code in
keep_select.pas, enhanced to read the input a field at a time instead
of in large binary chunks, thus avoiding the problem caused by our
changing all 'extended' fields to 'double'.